Update fetch.js
Loopback returns `error`, which is an object with the `message` key inside. 

I was about to code it directly as `json.error.message` but I was not sure if the server also sends `json.message` for some status codes. So I decided to keep both with an OR. Feel free to change this.